move: infractions commands to infractions module and category
Closes #36
