+++ src/commands.rs
@@ -1,3 +1,4 @@
+pub mod infractions;
 pub mod misc;
 pub mod moderation;
 pub mod utility;
